ensure file uses unix line feeds! (#8)
While I was in the right place, unbalanced quotes was symptom not cause.
The DOS line feeds were breaking the shell var used for the awk program.

Also the single line awk program converted to multiline for readability.
I'm not  familiar with the data or awk so can't guarantee results are
identical; local tests are successful.

Useful resource for understanding awk:
http://www.tutorialspoint.com/awk/awk_basic_syntax.htm